pyproject.toml
[project]
name = "shoppoint"
version = "0.1.0"
description = "FastAPI project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aio-pika>=9.5.4",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.6",
    "fastapi-pagination>=0.12.34",
    "mypy>=1.14.1",
    "pydantic-settings>=2.7.1",
    "pydantic[email]>=2.10.5",
    "pyjwt[crypto]>=2.10.1",
    "python-multipart>=0.0.20",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.37",
    "uvicorn>=0.34.0",
    "loguru>=0.7.3",
    "pydantic-extra-types>=2.10.2",
    "bcrypt>=4.2.1",
    "sqlakeyset>=2.0.1733540710",
    "python-magic-bin>=0.4.14",
    "aiofiles>=24.1.0",
    "aioboto3>=13.4.0",
]

[project.optional-dependencies]
cli = [
    "ruff>=0.9.1",
    "bandit>=1.8.2",
]

[tool.ruff]
exclude = ["**/__init__.py"]

